Real-Estate-Project/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-solo==2.1.0
Pillow==10.1.0
pip==23.2.1
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.3.1